requirements.txt
wcwidth
PySide6-Essentials